# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "KAI API",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/stylist/add_salon": {
            "post": {
                "tags": [
                    "Salon"
                ],
                "summary": "create salon",
                "description": "create salon and if exist , then you can update it using this API",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "salon's name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "location",
                        "in": "path",
                        "description": "location",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "latitude",
                        "in": "path",
                        "description": "latitude",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "longitude",
                        "in": "path",
                        "description": "longitude",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "images[]",
                        "in": "path",
                        "description": "images max 10 pieces",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation message "
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/show_my_salon": {
            "get": {
                "tags": [
                    "Salon"
                ],
                "summary": "show salon info",
                "description": "get salon info like name,images ,address for owner 'stylist'",
                "responses": {
                    "200": {
                        "description": "salon full info and images",
                        "schema": {
                            "$ref": "#/definitions/Salon"
                        }
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/update_location": {
            "post": {
                "tags": [
                    "Salon"
                ],
                "summary": "update salon's location",
                "description": "update salon location for salon owner 'stylist'",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "description": "location will be updated by loation object that consist ,address, latitude,longitude ",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "location update",
                        "schema": {
                            "$ref": "#/definitions/Salon"
                        }
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/show_my_location": {
            "get": {
                "tags": [
                    "Salon"
                ],
                "summary": "show salon's location",
                "description": "get salon location for salon owner 'stylist'",
                "responses": {
                    "200": {
                        "description": "return location and iside lat and long",
                        "schema": {
                            "$ref": "#/definitions/Salon"
                        }
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/add_service": {
            "post": {
                "tags": [
                    "Salon"
                ],
                "summary": "add service to salon",
                "description": "add services",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "path",
                        "description": "price",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "service added successfuly",
                        "schema": {
                            "$ref": "#/definitions/Services"
                        }
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/my_service": {
            "get": {
                "tags": [
                    "Salon"
                ],
                "summary": "show salon's services for stylis 'for self'",
                "description": "get salon services'",
                "responses": {
                    "200": {
                        "description": "return all salon service",
                        "schema": {
                            "$ref": "#/definitions/Services"
                        }
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/delete_service/{id}": {
            "get": {
                "tags": [
                    "Salon"
                ],
                "summary": "delete specifc service",
                "description": "delete sservice",
                "responses": {
                    "200": {
                        "description": "service deleted",
                        "schema": {
                            "$ref": "#/definitions/Services"
                        }
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Salon": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "images": {
                    "type": "text"
                },
                "latitude": {
                    "type": "double"
                },
                "longitude": {
                    "type": "double"
                }
            }
        },
        "Services": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                }
            }
        },
...
    }
}